Fix forward declaration signature mismatch

+++ fs/verity/signature.c
@@ -14,7 +14,7 @@
 
 /* Forward declaration */
 int __fsverity_verify_signature(const struct inode *inode, const u8 *signature,
-				size_t sig_size, const u8 *file_digest,
+				u32 sig_size, const u8 *file_digest,
 				unsigned int digest_algorithm);
 
 /*
